rescue mode can mount Linux partition again
Fixes "You don't have any Linux partitions." in the rescue mode.

It took me an hour to figure why did the rescue mode was not able to find
a Linux partition.

If we do not want to mount pre-existing partition nor detect the OS in
the installer, then we should only do that in the installer.

Then, it would be much more useful if the function isn't just getting
chopped off like it was but instead a message explaining why it was
chopped off added. Can save hours for people trying to figure why their
rescue mode isn't finding any Linux partition.