Reduce anaconda dependencies
Do not pull in a lot of network stuff, which we don't use. This saves
some space on the installation image, but also reduces dom0 size and
makes the menu entries a bit less confusing - since initial-setup-gui
would pull it all in into the installed system too.

Fixes QubesOS/qubes-issues#6894